rename module and remove openssl folder
