+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/golang-fips/openssl-fips
+module github.com/golang-fips/openssl
 
 go 1.18
