+++ goopenssl.c
