+++ hmac.go
