+++ hmac_test.go
