+++ rand.go
