+++ thread_setup.c
